Fix cast uint64_t to size_t

+++ picoquic/frames.c
@@ -2807,7 +2807,7 @@ void picoquic_process_possible_ack_of_ack_frame(picoquic_cnx_t* cnx, picoquic_pa
                     content_bytes = picoquic_decode_datagram_frame_header(content_bytes, content_bytes + frame_length,
                         &frame_id, &content_length);
 
-                    ret = (cnx->callback_fn)(cnx, 0, content_bytes, content_length,
+                    ret = (cnx->callback_fn)(cnx, 0, content_bytes, (size_t)content_length,
                         (is_spurious)? picoquic_callback_datagram_spurious:picoquic_callback_datagram_acked,
                         cnx->callback_ctx, NULL);
                 }
